+++ tests/TestCase/FilesystemTest.php
@@ -56,17 +56,17 @@ class FilesystemTest extends TestCase
     public function testCreateFile(): void
     {
         $filename = TMP . 'dirToBeCreated' . DS . 'exampleFile';
-        $this->assertTrue(Filesystem::instance()->createFile($filename));
+        $this->assertSame($filename, Filesystem::instance()->createFile($filename));
         $this->assertStringEqualsFile($filename, '');
 
         unlink($filename);
-        $this->assertTrue(Filesystem::instance()->createFile($filename, 'string'));
+        $this->assertSame($filename, Filesystem::instance()->createFile($filename, 'string'));
         $this->assertStringEqualsFile($filename, 'string');
 
         $this->skipIf(IS_WIN);
 
         //Using a no existing directory, but ignoring errors
-        $this->assertFalse(Filesystem::instance()->createFile(DS . 'noExistingDir' . DS . 'file', null, 0777, true));
+        $this->assertEmpty(Filesystem::instance()->createFile(DS . 'noExistingDir' . DS . 'file', null, 0777, true));
 
         //Using a no existing directory
         $this->expectException(IOException::class);